// src/modules/report/operational/report-operational.types.ts
export interface ReportFilter {
    start_date: string;
    end_date: string;
    shift_id?: string | null;
    user_id?: string | null;
}